query_system/requirements.txt

Django>=4.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2  
#psycopg2-binary>=2.9.9  # PostgreSQL (if using PostgreSQL)
pymongo>=4.5.0  
django-cors-headers>=4.3.1  
django-environ>=0.11.2  
pydantic>=2.0 
langchain>=0.1.0  
langchain-huggingface>=0.1.0  
gunicorn>=21.2.0
